requirements.txt
requests
numpy
tensorflow
keras
Pillow